🚧♻️ (lead) Button to manually send the email
