+++ som_leads_polissa/giscedata_crm_lead.py
@@ -2,6 +2,7 @@
 from datetime import datetime
 from osv import fields, osv
 import netsvc
+from oorq.decorators import job
 
 from base_extended_som.res_partner import GENDER_SELECTION
 
@@ -347,6 +348,44 @@ class GiscedataCrmLead(osv.OsvInherits):
         lead_id = super(GiscedataCrmLead, self).create(cursor, uid, vals, context=context)
 
         return lead_id
+
+    @job(queue="poweremail_>sender")
+    def _send_mail_async(self, cr, uid, lead_id, context=None):
+        self._send_mail(cr, uid, lead_id, context=context)
+
+    def _send_mail(self, cr, uid, lead_id, context=None):
+        if context is None:
+            context = {}
+
+        lead_o = self.pool.get("giscedata.crm.lead")
+        ir_model_o = self.pool.get('ir.model.data')
+        template_o = self.pool.get('poweremail.templates')
+
+        lead = lead_o.read(cr, uid, lead_id, ['create_new_member', 'polissa_id'], context=context)
+
+        template_name = "email_contracte_esborrany"
+        if lead["create_new_member"]:
+            template_name = "email_contracte_esborrany_nou_soci"
+        template_id = ir_model_o.get_object_reference(cr, uid, 'som_polissa_soci', template_name)[1]
+
+        polissa_id = lead["polissa_id"][0]
+        from_id = template_o.read(cr, uid, template_id)['enforce_from_account'][0]
+
+        wiz_send_obj = self.pool.get("poweremail.send.wizard")
+        context.update({
+            "active_ids": [polissa_id],
+            "active_id": polissa_id,
+            "template_id": template_id,
+            "src_model": "giscedata.polissa",
+            "src_rec_ids": [polissa_id],
+            "from": from_id,
+            "state": "single",
+            "priority": "0",
+        })
+
+        params = {"state": "single", "priority": "0", "from": context["from"]}
+        wiz_id = wiz_send_obj.create(cr, uid, params, context)
+        return wiz_send_obj.send_mail(cr, uid, [wiz_id], context)
 
     _columns = {
         "tipus_tarifa_lead": fields.selection(_tipus_tarifes_lead, "Tipus de tarifa del contracte"),
